CLI refactor: move scripts to euphonic.cli, rework entrypoint
- scripts moved to euphonic.cli
- shebangs and __main__ logic removed; pip installer handles this
- euphonic.script_utils is moved to euphonic.cli.utils
- scripts directory is removed
- main() function added to optimise_eta
- update commands in documentation
- tone down references to "scripts" in the docs; prefer "program" as
  these are no longer such approachable/editable files
