+++ tests_and_analysis/test/script_tests/regenerate_dos_data.py
@@ -1,6 +1,6 @@
 import json
 from unittest.mock import patch
-import scripts.dos
+import euphonic.cli.dos
 from tests_and_analysis.test.script_tests.utils import (
     get_phonon_file, get_dos_params, get_current_plot_lines_xydata,
     get_dos_data_file)
@@ -13,7 +13,7 @@ def regenerate_dos_data(_):
 
     for dos_params in get_dos_params():
         # Generate current figure for us to retrieve with gcf
-        scripts.dos.main([get_phonon_file()] + dos_params)
+        euphonic.cli.dos.main([get_phonon_file()] + dos_params)
 
         # Retrieve with gcf and record data
         json_data[" ".join(dos_params)] = get_current_plot_lines_xydata()
