+++ tests_and_analysis/test/script_tests/test_dos.py
@@ -6,7 +6,7 @@ import numpy.testing as npt
 # Required for mocking
 import matplotlib.pyplot
 from .utils import get_phonon_file, get_dos_params, get_dos_data_file
-import scripts.dos
+import euphonic.cli.dos
 
 
 @pytest.mark.integration
@@ -25,7 +25,7 @@ class TestRegression:
 
     @pytest.mark.parametrize("dos_args", get_dos_params())
     def test_plots_produce_expected_xydata(self, inject_mocks, dos_args):
-        scripts.dos.main([get_phonon_file()] + dos_args)
+        euphonic.cli.dos.main([get_phonon_file()] + dos_args)
 
         lines = matplotlib.pyplot.gcf().axes[0].lines
